requirements.txt
torch==2.*
torchvision==0.15.*
Pillow==9.*
numpy==1.*